% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSE.R
\name{createSE}
\alias{createSE}
\title{Title}
\usage{
createSE(dat.icb)
}
\arguments{
\item{dat.icb}{A MAE object including genomic, clinical and annotation data.}
}
\value{
A SE object.
}
\description{
Title
}
