% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CYT_Rooney.R
\docType{data}
\name{CYT_Rooney}
\alias{CYT_Rooney}
\title{CYT Rooney RNA Signature}
\format{
An object of class \code{data.frame} with 2 rows and 5 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/25594174/}
}
\usage{
CYT_Rooney
}
\description{
Gene expression signature capturing the local immune cytolytic (CYT) activity.
}
\keyword{datasets}
