% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMT_Thompson.R
\docType{data}
\name{EMT_Thompson}
\alias{EMT_Thompson}
\title{EMT Thompson RNA Signature}
\format{
An object of class \code{data.frame} with 12 rows and 5 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/31683225/}
}
\usage{
EMT_Thompson
}
\description{
Gene expression signature capturing the epithelial-to-mesenchymal transition (EMT) activity. This EMT signature was a predictor of the failure of PD-(L)1 blockade in lung cancer patients.
}
\keyword{datasets}
