% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlotPerTreatment.R
\name{forestPlotPerTreatment}
\alias{forestPlotPerTreatment}
\title{Title}
\usage{
forestPlotPerTreatment(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  feature,
  xlab,
  label
)
}
\arguments{
\item{coef}{}

\item{se}{}

\item{study}{}

\item{pval}{}

\item{n}{}

\item{cancer.type}{}

\item{treatment}{}

\item{feature}{}

\item{xlab}{}

\item{label}{}
}
\description{
Title
}
