% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLogReg.R
\name{geneLogReg}
\alias{geneLogReg}
\title{Title}
\usage{
geneLogReg(
  dat.icb,
  clin = NULL,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  feature,
  study,
  n0.cutoff,
  n1.cutoff,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{aaaaa}

\item{clin}{bbbbbb}

\item{missing.perc}{ccccc}

\item{const.int}{ddddd}

\item{n.cutoff}{eeee}

\item{feature}{fffff}

\item{study}{gggg}

\item{n0.cutoff}{hhhhh}

\item{n1.cutoff}{iiii}

\item{cancer.type}{jjjj}

\item{treatment}{kkkkk}
}
\value{
lllll
}
\description{
Title
}
