% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigCOX_IS.R
\name{geneSigCOX_IS}
\alias{geneSigCOX_IS}
\title{Title}
\usage{
geneSigCOX_IS(
  dat.icb,
  clin = NULL,
  sig,
  sig.name,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  sig.perc,
  study
)
}
\arguments{
\item{dat.icb}{}

\item{clin}{}

\item{sig}{}

\item{sig.name}{}

\item{missing.perc}{}

\item{const.int}{}

\item{n.cutoff}{}

\item{sig.perc}{}

\item{study}{}
}
\description{
Title
}
