% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigIPRES.R
\name{geneSigIPRES}
\alias{geneSigIPRES}
\title{Compute IPRES Signature Score}
\usage{
geneSigIPRES(
  dat.icb,
  clin = NULL,
  sig,
  sig.name,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  sig.perc,
  study
)
}
\arguments{
\item{dat.icb}{A MultiAssayExperiment (MAE) object, SummarizedExperiment (SE) object, or a data frame or matrix of gene expression data.}

\item{clin}{If dat.icb is a data frame or matrix, then it contains clinical data (as data frame or matrix). By default, it is NULL.}

\item{sig}{A data frame of list of genes' symbol named 'gene_name'.}

\item{sig.name}{Name of signature.}

\item{missing.perc}{A cutoff to remove genes with zero expression across samples.}

\item{const.int}{A pseudocount is added to the TPM (Transcripts Per Million) values before performing a log transformation.}

\item{n.cutoff}{Minimum number of samples included in the association analysis.}

\item{sig.perc}{Minimum percentage of genes in a given expression data.}

\item{study}{Name of study.}
}
\value{
A numeric vector of computed signature score.
}
\description{
Consider the specific method explained at ... to compute IPRES signature score.
}
\examples{
Compute the signature score for IPRES Hugo signature using the specific method. 
geneSigIPRES(dat.icb = ICB_small_Liu, 
             sig = IPRES_Hugo,
             sig.name = 'IPRES_Hugo',
             missing.perc = 0.5,
             const.int = 0.001,
             n.cutoff = 15,
             sig.perc = 0.8, 
             study = 'ICB_Liu')
            
}
