% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigIPS.R
\name{geneSigIPS}
\alias{geneSigIPS}
\title{Compute IPS Signature Score}
\usage{
geneSigIPS(
  dat.icb,
  sig,
  sig.name,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  study,
  gene.annot = "gene_name"
)
}
\arguments{
\item{dat.icb}{A MultiAssayExperiment (MAE) object, SummarizedExperiment (SE) object, or a data frame or matrix of gene expression data.}

\item{sig}{A data frame of list of genes' symbol named 'gene_name'.}

\item{sig.name}{Name of signature.}

\item{missing.perc}{A cutoff to remove genes with zero expression across samples.}

\item{const.int}{A pseudocount is added to the TPM (Transcripts Per Million) values before performing a log transformation.}

\item{n.cutoff}{Minimum number of samples included in the association analysis.}

\item{study}{Name of study.}
}
\value{
A numeric vector of computed signature score.
}
\description{
Consider the specific method explained at ... to compute IPS signature score.
}
\examples{
Compute the signature score for IPS Charoentong signature using the specific method. 
geneSigIPS(dat.icb = ICB_small_Mariathasan, 
           sig = IPS_Charoentong,
           sig.name = 'IPS_Charoentong',
           missing.perc = 0.5,
           const.int = 0.001,
           n.cutoff = 15,
           study = 'ICB_Mariathasan')
             
}
