% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigLogReg.R
\name{geneSigLogReg}
\alias{geneSigLogReg}
\title{Title}
\usage{
geneSigLogReg(
  dat.icb,
  clin = NULL,
  geneSig,
  n.cutoff,
  study,
  sig.name,
  n0.cutoff,
  n1.cutoff,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{aaaaa}

\item{clin}{bbbbb}

\item{geneSig}{ccccc}

\item{n.cutoff}{ddddd}

\item{study}{eeeee}

\item{sig.name}{fffff}

\item{n0.cutoff}{ggggg}

\item{n1.cutoff}{hhhhh}

\item{cancer.type}{iiiii}

\item{treatment}{jjjjjj}
}
\value{
kkkkk
}
\description{
Title
}
