% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigMean.R
\name{geneSigMean}
\alias{geneSigMean}
\title{Compute Signature Score using Weighted Mean Average}
\usage{
geneSigMean(
  dat.icb,
  sig,
  sig.name,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  sig.perc,
  study
)
}
\arguments{
\item{dat.icb}{A MultiAssayExperiment (MAE) object, SummarizedExperiment (SE) object, or a data frame or matrix of gene expression data.}

\item{sig}{A data frame of list of genes' symbol named 'gene_name'.}

\item{sig.name}{Name of signature.}

\item{missing.perc}{A cutoff to remove genes with zero expression across samples.}

\item{const.int}{A pseudocount is added to the TPM (Transcripts Per Million) values before performing a log transformation.}

\item{n.cutoff}{Minimum number of samples included in the association analysis.}

\item{sig.perc}{Minimum percentage of genes in a given expression data.}

\item{study}{Name of study.}
}
\value{
A numeric vector of computed signature score.
}
\description{
Consider the weighted mean average to compute signature score.
}
\examples{
Compute the signature score for EMT Thompson signature using the weighted mean method. 
geneSigMean(dat.icb = ICB_small_Mariathasan, 
            sig = EMT_Thompson,
            sig.name = 'EMT_Thompson',
            missing.perc = 0.5,
            const.int = 0.001,
            n.cutoff = 15,
            sig.perc = 0.8, 
            study = 'ICB_Mariathasan')
             
}
