% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigPassON.R
\name{geneSigPassON}
\alias{geneSigPassON}
\title{Compute PassON Signature Score}
\usage{
geneSigPassON(
  dat.icb,
  sig,
  sig.name,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  sig.perc,
  study,
  gene.annot = "gene_name"
)
}
\arguments{
\item{dat.icb}{A MultiAssayExperiment (MAE) object, SummarizedExperiment (SE) object, or a data frame or matrix of gene expression data.}

\item{sig}{A data frame of list of genes' symbol named 'gene_name'.}

\item{sig.name}{Name of signature.}

\item{missing.perc}{A cutoff to remove genes with zero expression across samples.}

\item{const.int}{A pseudocount is added to the TPM (Transcripts Per Million) values before performing a log transformation.}

\item{n.cutoff}{Minimum number of samples included in the association analysis.}

\item{sig.perc}{Minimum percentage of genes in a given expression data.}

\item{study}{Name of study.}

\item{gene.annot}{Specify gene annotation including gene symbol (i.e., gene_name), ENTREZ ID (i.e., entrez_id), and ENSEMBL gene ID (i.e., gene_id).}
}
\value{
A numeric vector of computed signature score.
}
\description{
Consider the specific method explained at ... to compute PassON signature score.
}
\examples{
Compute the signature score for PassON Du signature using the specific method. 
geneSigPassON(dat.icb = ICB_small_Mariathasan, 
              sig = PassON_Du,
              sig.name = 'PassON_Du',
              missing.perc = 0.5,
              const.int = 0.001,
              n.cutoff = 15,
              sig.perc = 0.8, 
              study = 'ICB_Mariathasan')
             
}
