% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigSurvCont.R
\name{geneSigSurvCont}
\alias{geneSigSurvCont}
\title{Fit Proportional Hazards Regression Model for Gene Signature: Continuous Variable}
\usage{
geneSigSurvCont(
  dat.icb,
  clin = NULL,
  geneSig,
  time.censor,
  n.cutoff,
  study,
  surv.outcome,
  sig.name,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{A MultiAssayExperiment (MAE) object, SummarizedExperiment (SE) object, or a data frame or matrix of gene expression data.}

\item{clin}{If dat.icb is a data frame or matrix, then it contains clinical data (as data frame or matrix). By default, it is NULL.}

\item{geneSig}{A numeric vector of computed signature score.}

\item{time.censor}{Possible censoring in months.}

\item{n.cutoff}{Minimum number of samples included in the association analysis.}

\item{study}{Name of study.}

\item{surv.outcome}{Overall survival (OS) or progression-free survival (PFS).}

\item{sig.name}{Name of signature.}

\item{cancer.type}{Name of the cancer type for the given study.}

\item{treatment}{Name of the treatment for the given study.}
}
\description{
Fits a Cox proportional hazards regression model with continuous signature data using the counting process formulation of Andersen and Gill.
}
