% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigSurvDicho.R
\name{geneSigSurvDicho}
\alias{geneSigSurvDicho}
\title{Title}
\usage{
geneSigSurvDicho(
  dat.icb,
  clin = NULL,
  geneSig,
  time.censor,
  n.cutoff,
  n0.cutoff,
  n1.cutoff,
  study,
  surv.outcome,
  sig.name,
  method = "median",
  var.type,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{aaaaa}

\item{clin}{bbbbb}

\item{geneSig}{ccccc}

\item{time.censor}{ddddd}

\item{n.cutoff}{eeeee}

\item{n0.cutoff}{fffff}

\item{n1.cutoff}{ggggg}

\item{study}{hhhhh}

\item{surv.outcome}{iiii}

\item{sig.name}{jjjjj}

\item{method}{kkkkk}

\item{var.type}{lllll}

\item{cancer.type}{mmmm}

\item{treatment}{nnnnn}
}
\value{
ooooo
}
\description{
Title
}
