% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSurvCont.R
\name{geneSurvCont}
\alias{geneSurvCont}
\title{Title}
\usage{
geneSurvCont(
  dat.icb,
  clin = NULL,
  time.censor,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  feature,
  study,
  surv.outcome,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{aaaa}

\item{clin}{bbbb}

\item{time.censor}{cccc}

\item{missing.perc}{ddddd}

\item{const.int}{eeee}

\item{n.cutoff}{fffff}

\item{feature}{ggggg}

\item{study}{hhhhh}

\item{surv.outcome}{iiii}

\item{cancer.type}{jjjjj}

\item{treatment}{kkkkk}
}
\value{
lllll
}
\description{
Title
}
