% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_Liu.R
\docType{data}
\name{ICB_Liu}
\alias{ICB_Liu}
\title{ICB Liu Melanoma}
\format{
An object of class \code{SummarizedExperiment} with 16737 rows and 74 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/31792460/}
}
\usage{
ICB_Liu
}
\description{
Immune checkpoint blockade (ICB) Liu Melnaoma immunotherapy (PD-1) dataset.
}
\details{
A SE object with 16,737 protein-coding genes and 74 patients were treated immunotherapy targeted PD-1:
}
\keyword{datasets}
