% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_Mariathasan.R
\docType{data}
\name{ICB_Mariathasan}
\alias{ICB_Mariathasan}
\title{ICB Mariathasan Bladder}
\format{
An object of class \code{SummarizedExperiment} with 17993 rows and 195 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29443960/}
}
\usage{
ICB_Mariathasan
}
\description{
Immune checkpoint blockade (ICB) Mariathasan Bladder immunotherapy (PD-1/PD-L1) dataset. A SummarizedExperiment object with 17,993 protein-coding genes and 195 patients was treated with immunotherapy targeting PD-1/PD-L1.
}
\keyword{datasets}
