% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_Miao.R
\docType{data}
\name{ICB_Miao}
\alias{ICB_Miao}
\title{ICB Miao Kidney}
\format{
An object of class \code{SummarizedExperiment} with 20242 rows and 33 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29301960/}
}
\usage{
ICB_Miao
}
\description{
Immune checkpoint blockade (ICB) Miao Kidney immunotherapy (PD-1) dataset. A SummarizedExperiment object with 20,242 protein-coding genes and 33 patients was treated with immunotherapy targeting PD-1.
}
\keyword{datasets}
