% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_small_Hugo.R
\docType{data}
\name{ICB_small_Hugo}
\alias{ICB_small_Hugo}
\title{ICB Hugo Melanoma}
\format{
An object of class \code{SummarizedExperiment} with 1801 rows and 27 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/26997480/}
}
\usage{
ICB_small_Hugo
}
\description{
Immune checkpoint blockade (ICB) Hugo Melnaoma immunotherapy (PD-1) dataset. A subset of SummarizedExperiment object with 19,961 protein-coding genes and 27 patients was treated with immunotherapy targeting PD-1.
}
\keyword{datasets}
