% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_small_Liu.R
\docType{data}
\name{ICB_small_Liu}
\alias{ICB_small_Liu}
\title{ICB Liu Melanoma}
\format{
An object of class \code{SummarizedExperiment} with 1741 rows and 74 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/31792460/}
}
\usage{
ICB_small_Liu
}
\description{
Immune checkpoint blockade (ICB) Liu Melnaoma immunotherapy (PD-1) dataset. A subset of SummarizedExperiment object with 16,737 protein-coding genes and 74 patients was treated with immunotherapy targeting PD-1.
}
\keyword{datasets}
