% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_small_Nathanson.R
\docType{data}
\name{ICB_small_Nathanson}
\alias{ICB_small_Nathanson}
\title{ICB Nathanson Melanoma}
\format{
An object of class \code{SummarizedExperiment} with 1792 rows and 24 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/27956380/}
}
\usage{
ICB_small_Nathanson
}
\description{
Immune checkpoint blockade (ICB) Nathanson Melnaoma immunotherapy (CTLA-4) dataset. A subset of SummarizedExperiment object with 19,279 protein-coding genes and 24 patients was treated with immunotherapy targeting CTLA-4.
}
\keyword{datasets}
