% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_small_Padron.R
\docType{data}
\name{ICB_small_Padron}
\alias{ICB_small_Padron}
\title{ICB Padron Pancreas}
\format{
An object of class \code{SummarizedExperiment} with 1792 rows and 45 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/35662283/}
}
\usage{
ICB_small_Padron
}
\description{
Immune checkpoint blockade (ICB) Padron Bladder immunotherapy (PD-1) dataset. A subset of SummarizedExperiment object with 18,459 protein-coding genes and 45 patients was treated with immunotherapy targeting PD-1.
}
\keyword{datasets}
