% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_small_Riaz.R
\docType{data}
\name{ICB_small_Riaz}
\alias{ICB_small_Riaz}
\title{ICB Riaz Melanoma}
\format{
An object of class \code{SummarizedExperiment} with 1801 rows and 23 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29033130/}
}
\usage{
ICB_small_Riaz
}
\description{
Immune checkpoint blockade (ICB) Riaz Melanoma immunotherapy (PD-1/PD-L1 & CTLA-4) dataset. A subset of SummarizedExperiment object with 19,961 protein-coding genes and 23 patients was treated with immunotherapy targeting PD-1/PD-L1 & CTLA-4
}
\keyword{datasets}
