% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICB_Van_Allen.R
\docType{data}
\name{ICB_Van_Allen}
\alias{ICB_Van_Allen}
\title{ICB Van Allen Melanoma}
\format{
An object of class \code{SummarizedExperiment} with 20242 rows and 42 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/26359337/}
}
\usage{
ICB_Van_Allen
}
\description{
Immune checkpoint blockade (ICB) Van Allen Melnaoma immunotherapy (CTLA-4) dataset. A SummarizedExperiment object with 20,242 protein-coding genes and 42 patients was treated with immunotherapy targeting CTLA-4.
}
\keyword{datasets}
