% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPRES_Hugo.R
\docType{data}
\name{IPRES_Hugo}
\alias{IPRES_Hugo}
\title{IPRES Hugo RNA Signature}
\format{
An object of class \code{list} of length 39.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/26997480/}
}
\usage{
IPRES_Hugo
}
\description{
Gene expression signature capturing the combined epithelial-to-mesenchymal transition (EMT) and Inflammation activities. This EMT/Inflammation signature, called innate anti-PD1 resistance (IPRES) was a predictor of the failure of PD-(L)1 blockade in melanoma patients.
}
\keyword{datasets}
