% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS_Charoentong.R
\docType{data}
\name{IPS_Charoentong}
\alias{IPS_Charoentong}
\title{IPS Charoentong RNA Signature}
\format{
An object of class \code{data.frame} with 162 rows and 7 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/28052254/}
}
\usage{
IPS_Charoentong
}
\description{
Using machine learning, the authors identified determinants of tumor immunogenicity and developed a scoring scheme for the quantification termed immunophenoscore (IPS). This IPS signature was a predictor of the response of CTLA-4 and PD-1 in published cohorts.
}
\keyword{datasets}
