% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSOV_Shen.R
\docType{data}
\name{IPSOV_Shen}
\alias{IPSOV_Shen}
\title{IPSOV Shen RNA Signature}
\format{
An object of class \code{list} of length 15.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/30594555/}
}
\usage{
IPSOV_Shen
}
\description{
Gene expression signature from immune genes to develop an immune-based prognostic score for OV (IPSOV). IPSOV is a promising prognostic biomarker in ovarian cancer which could be used to distinguish and predict patients' survival outcome.
}
\keyword{datasets}
