% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMPlot.R
\name{KMPlot}
\alias{KMPlot}
\title{Title}
\usage{
KMPlot(
  surv,
  time,
  time.censor,
  var,
  title,
  xlab,
  ylab,
  method = "median",
  n0.cutoff,
  n1.cutoff,
  var.type = TRUE
)
}
\arguments{
\item{surv}{aaaa}

\item{time}{bbbbb}

\item{time.censor}{ccccc}

\item{var}{dddd}

\item{title}{eeee}

\item{xlab}{fffff}

\item{ylab}{gggg}

\item{method}{hhhh}

\item{n0.cutoff}{iiii}

\item{n1.cutoff}{jjjjj}

\item{var.type}{kkkkk}
}
\value{
llll
}
\description{
Title
}
