% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M1_Hwang.R
\docType{data}
\name{M1_Hwang}
\alias{M1_Hwang}
\title{M1 Hwang RNA Signature}
\format{
An object of class \code{data.frame} with 13 rows and 5 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/31959763/}
}
\usage{
M1_Hwang
}
\description{
Gene expression signature capturing the presence Macrophage M1 cells. This M1 signature was a predictor of the response of PD-1 blockade in melanoma patients.
}
\keyword{datasets}
