% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafun.R
\name{metafun}
\alias{metafun}
\title{Title}
\usage{
metafun(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  cancer.spec = FALSE,
  treatment.spec = FALSE,
  feature
)
}
\arguments{
\item{coef}{}

\item{se}{}

\item{study}{}

\item{pval}{}

\item{n}{}

\item{cancer.type}{}

\item{treatment}{}

\item{cancer.spec}{}

\item{treatment.spec}{}

\item{feature}{}
}
\description{
Title
}
