% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPerCanfun.R
\name{metaPerCanfun}
\alias{metaPerCanfun}
\title{Title}
\usage{
metaPerCanfun(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  cancer.spec = TRUE,
  feature
)
}
\arguments{
\item{coef}{}

\item{se}{}

\item{study}{}

\item{pval}{}

\item{n}{}

\item{cancer.type}{}

\item{treatment}{}

\item{cancer.spec}{}

\item{feature}{}
}
\description{
Title
}
