% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PassON_Du.R
\docType{data}
\name{PassON_Du}
\alias{PassON_Du}
\title{PassON Du RNA Signature}
\format{
An object of class \code{list} of length 4.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/34654806/}
}
\usage{
PassON_Du
}
\description{
Gene expression signature capturing the combined fatty acid and lipid metabolism-related signatures in terms of response to PD1/PD-L1 therapies in patients with metastatic melanoma.
}
\keyword{datasets}
