% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictIO_Bareche.R
\docType{data}
\name{PredictIO_Bareche}
\alias{PredictIO_Bareche}
\title{PredictIO Bareche RNA Signature}
\format{
An object of class \code{data.frame} with 100 rows and 6 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/36055464/}
}
\usage{
PredictIO_Bareche
}
\description{
Gene expression signature that demonstrated better and more consistent ability to predict immunotherapy response as compared to other signatures. The specific algorithm is applied to computed the signature score. The PredictIO signature was identified by integrating gene expression data from a large pan-cancer discovery immunotherapy datasets.
}
\keyword{datasets}
