% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{rem}
\alias{rem}
\title{Title}
\usage{
rem(x, missing.perc, const.int)
}
\arguments{
\item{x}{An expression data matrix.}

\item{missing.perc}{A cutoff for percentage of missing variables across samples.}

\item{const.int}{A constant value to do log-transformed using TPM data.}
}
\value{
An expression data after removing low-expressed genes.
}
\description{
Title
}
