% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalefun.R
\name{scalefun}
\alias{scalefun}
\title{Scaling and Centering the Matrix of Expression Data}
\usage{
scalefun(x)
}
\arguments{
\item{x}{A numeric matrix of expression data.}
}
\value{
A scaled matrix.
}
\description{
A generic function whose default method centers and scales the rows (or features) of a numeric expression matrix.
}
