% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCont.R
\name{survCont}
\alias{survCont}
\title{Title}
\usage{
survCont(surv, time, time.censor, var)
}
\arguments{
\item{surv}{aaaa}

\item{time}{bbbb}

\item{time.censor}{cccc}

\item{var}{dddd}
}
\value{
hhhh
}
\description{
Title
}
