% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDicho.R
\name{survDicho}
\alias{survDicho}
\title{Title}
\usage{
survDicho(
  surv,
  time,
  time.censor,
  var,
  n0.cutoff,
  n1.cutoff,
  method = "median",
  var.type = TRUE
)
}
\arguments{
\item{surv}{aaaaa}

\item{time}{bbbbb}

\item{time.censor}{ccccc}

\item{var}{dddd}

\item{n0.cutoff}{eeee}

\item{n1.cutoff}{ffff}

\item{method}{gggg}

\item{var.type}{hhhhh}
}
\value{
iiiii
}
\description{
Title
}
