% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano Plot}
\usage{
volcanoPlot(
  feature,
  coef,
  pval,
  padj,
  pos.cutoff,
  neg.cutoff,
  x.lab,
  padj.label,
  cutoff,
  colors,
  coef.cutoff
)
}
\arguments{
\item{feature}{A vector of characters showing the features' names.}

\item{coef}{A numeric vector of log-odds or log-hazard ratio estimates.}

\item{pval}{A numeric vector of estimated p-values.}

\item{padj}{A numeric vector of adjusted p-values to correct for multiple tests using the false discovery rate (FDR).}

\item{pos.cutoff}{Number of selected top features with positive log-odds or log-hazard ratio estimates.}

\item{neg.cutoff}{Number of selected top features with negative log-odds or log-hazard ratio estimates.}

\item{x.lab}{Label of x-axis.}

\item{padj.label}{If the adjusted p-values (FDR) are used to select significantly associated features, then it is TRUE.}

\item{cutoff}{Cut-off for adjusted p-values (FDR) or p-values.}

\item{colors}{A vector of colors to identify positive, negative, and not associated features.}

\item{coef.cutoff}{Cut-off for estimated coef i.e., log-odds or log-hazard ratio.}
}
\description{
Creates a volcano plot of log-odds or log-hazard ratios of changes versus the log p-values from association analysis.
}
