% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COX_IS_Bonavita.R
\docType{data}
\name{COX_IS_Bonavita}
\alias{COX_IS_Bonavita}
\title{COX-IS Bonavita RNA Signature}
\format{
An object of class \code{data.frame} with 24 rows and 6 columns.
}
\source{
https://doi.org/10.1016/j.immuni.2020.10.020
}
\usage{
COX_IS_Bonavita
}
\description{
Gene expression signature capturing the pro-tumoral COX Inflammatory activity. This COX Inflammatory Signature (COX-IS) was generated in mouse data and found a predictor of the failure of CTLA-4 and PD-(L)1 blockade in several published cohorts.
}
\keyword{datasets}
