% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSE.R
\name{createSE}
\alias{createSE}
\title{Convert MultiAssayExperiment Object into SummarizedExperiment Object}
\usage{
createSE(dat.icb)
}
\arguments{
\item{dat.icb}{The MAE object including RNA, DNA, clinical and annotation data.}
}
\value{
A SummarizedExperiment object.
}
\description{
Convert MultiAssayExperiment (MAE) object into SummarizedExperiment (SE) object which is limited to protein-coding genes.
}
