% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlot.R
\name{forestPlot}
\alias{forestPlot}
\title{Title}
\usage{
forestPlot(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  xlab,
  label,
  feature
)
}
\arguments{
\item{coef}{}

\item{se}{}

\item{study}{}

\item{pval}{}

\item{n}{}

\item{cancer.type}{}

\item{treatment}{}

\item{xlab}{}

\item{label}{}

\item{feature}{}
}
\description{
Title
}
