% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlotPerTreatment.R
\name{forestPlotPerTreatment}
\alias{forestPlotPerTreatment}
\title{Forest Plot: Treatment-specific Analysis}
\usage{
forestPlotPerTreatment(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  feature,
  xlab,
  label
)
}
\arguments{
\item{coef}{A numeric vector of estimates of treatment effect, e.g., log hazard ratio or log odds ratio.}

\item{se}{A numeric vector of standard errors of treatment estimate.}

\item{study}{A vector of characters showing the studies' names.}

\item{pval}{A numeric vector of estimated p-values.}

\item{n}{A numeric vector of number of samples.}

\item{cancer.type}{A vector of characters showing the cancer types.}

\item{treatment}{A vector of characters showing the treatment types.}

\item{feature}{Name of feature.}

\item{xlab}{A label for x-axis.}

\item{label}{A label for estimates of treatment effect, e.g., log hazard ratio (logHR).}
}
\description{
Draws a forest plot in the active graphics window (using grid graphics system) across treatment types.
}
\examples{
expr <- list('ICB_Liu' = ICB_small_Liu, 'ICB_Padron' = ICB_small_Padron, 'ICB_Hugo' = ICB_small_Hugo, 
             'ICB_Mariathasan' = ICB_small_Mariathasan, 'ICB_Nathanson' = ICB_small_Nathanson, 
             'ICB_Riaz' = ICB_small_Riaz, 'ICB_Miao' = ICB_small_Miao, 'ICB_Van_Allen' = ICB_small_Van_Allen)

cancer_type <- c('Melanoma', 'Pancreas', 'Melanoma', 'Bladder', 'Melanoma', 'Melanoma', 'Kidney', 'Melanoma')
treatment_type <- c('PD-1/PD-L1', 'PD-1/PD-L1', 'PD-1/PD-L1', 'PD-1/PD-L1', 'CTLA4', 'IO+combo', 'PD-1/PD-L1', 'CTLA4')

assoc.res <- lapply(1:length(expr), function(k){

 geneSurvCont(dat.icb = expr[[k]],
 time.censor = 36,
 missing.perc = 0.5,
 const.int = 0.001,
 n.cutoff = 15,
 feature = 'CXCL9',
 study = names(expr)[k],
 surv.outcome = 'OS',
 cancer.type = cancer_type[k],
 treatment = treatment_type[k])
 
})
assoc.res <- do.call(rbind, assoc.res)

forestPlotPerTreatment(coef = assoc.res$Coef, 
                       se = assoc.res$SE,
                       study = assoc.res$Study,
                       pval = assoc.res$Pval,
                       n = assoc.res$N,
                       cancer.type = assoc.res$Cancer_type,
                       treatment = assoc.res$Treatment,
                       xlab = 'logOR estimate',
                       label = 'logOR',
                       feature = "CXCL9")
           
}
