% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestPlotPerTreatment.R
\name{forestPlotPerTreatment}
\alias{forestPlotPerTreatment}
\title{Forest Plot: Treatment-specific Analysis}
\usage{
forestPlotPerTreatment(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  feature,
  xlab,
  label
)
}
\arguments{
\item{coef}{A numeric vector of estimates of treatment effect, e.g., log hazard ratio or log odds ratio.}

\item{se}{A numeric vector of standard errors of treatment estimate.}

\item{study}{A vector of characters showing the studies' names.}

\item{pval}{A numeric vector of estimated p-values.}

\item{n}{A numeric vector of number of samples.}

\item{cancer.type}{A vector of characters showing the cancer types.}

\item{treatment}{A vector of characters showing the treatment types.}

\item{feature}{Name of feature.}

\item{xlab}{A label for x-axis.}

\item{label}{A label for estimates of treatment effect, e.g., log hazard ratio (logHR).}
}
\description{
Draws a forest plot in the active graphics window (using grid graphics system) across treatment types.
}
