% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLogReg.R
\name{geneLogReg}
\alias{geneLogReg}
\title{Fit Logistic Regression Model for Genes: Continuous Expression Variable}
\usage{
geneLogReg(
  dat.icb,
  clin = NULL,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  feature,
  study,
  n0.cutoff,
  n1.cutoff,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{A MultiAssayExperiment (MAE) object, SummarizedExperiment (SE) object, or a data frame or matrix of gene expression data.}

\item{clin}{If dat.icb is a data frame or matrix, then it contains clinical data (as data frame or matrix). By default, it is NULL.}

\item{missing.perc}{A cutoff to remove genes with zero expression across samples.}

\item{const.int}{A pseudocount is added to the TPM (Transcripts Per Million) values before performing a log transformation.}

\item{n.cutoff}{Minimum number of samples included in the association analysis.}

\item{feature}{A vector of character strings for selected features.}

\item{study}{Name of study.}

\item{n0.cutoff}{Minimum number of non-responders in the analysis.}

\item{n1.cutoff}{Minimum number of responders in the analysis.}

\item{cancer.type}{Name of the cancer type for the given study.}

\item{treatment}{Name of the treatment for the given study.}

\item{time.censor}{Possible censoring in months.}
}
\value{
A subset of results using an object of class logistic regression representing the fit.
Outcome: Immunotherapy response outcome i.e., R (responder) and NR (non-responder).
Gene: Name of selected genes.
Study: Name of study.
Coef: Estimate of treatment effect i.e., log odds ratio.
SE: Standard error of treatment estimate.
N: Number of samples.
Pval: Estimated p-value.
Cancer_type: A character shows the cancer type.
Treatment: A character shows the treatment type.
}
\description{
Fits a logistic regression model with continuous expression data for each gene.
}
\examples{
Assess the association between selected genes and response (R vs NR) in immunotherapy.
geneLogReg(dat.icb = ICB_Liu,
           missing.perc = 0.5,
           const.int = 0.001,
           n.cutoff = 15,
           feature = c('CXCL9', 'CXCL10', 'TIGIT', 
                         'CD83', 'STAT1', 'CXCL11',
                         'CXCL13', 'CD8A', 'CTLA4'),
           study = 'ICB_Liu', 
           n0.cutoff = 10,
           n1.cutoff = 10,
           cancer.type = 'Melanoma',
           treatment = 'PD-1/PD-L1')

}
