% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigSurvCont.R
\name{geneSigSurvCont}
\alias{geneSigSurvCont}
\title{Title}
\usage{
geneSigSurvCont(
  dat.icb,
  clin = NULL,
  geneSig,
  time.censor,
  n.cutoff,
  study,
  surv.outcome,
  sig.name,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{aaaaa}

\item{clin}{bbbbb}

\item{geneSig}{cccccc}

\item{time.censor}{dddddd}

\item{n.cutoff}{eeeee}

\item{study}{fffff}

\item{surv.outcome}{ggggg}

\item{sig.name}{hhhhh}

\item{cancer.type}{iiiii}

\item{treatment}{jjjjj}
}
\value{
kkkkk
}
\description{
Title
}
