% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSigSurvDicho.R
\name{geneSigSurvDicho}
\alias{geneSigSurvDicho}
\title{Fit Proportional Hazards Regression Model for Gene Signature: Dichotomous Variable}
\usage{
geneSigSurvDicho(
  dat.icb,
  clin = NULL,
  geneSig,
  time.censor,
  n.cutoff,
  n0.cutoff,
  n1.cutoff,
  study,
  surv.outcome,
  sig.name,
  method = "median",
  var.type,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{A MultiAssayExperiment (MAE) object, SummarizedExperiment (SE) object, or a data frame or matrix of gene expression data.}

\item{clin}{If dat.icb is a data frame or matrix, then it contains clinical data (as data frame or matrix). By default, it is NULL.}

\item{geneSig}{A numeric vector of computed signature score.}

\item{time.censor}{Possible censoring in months.}

\item{n.cutoff}{Minimum number of samples included in the association analysis.}

\item{n0.cutoff}{Minimum number of samples with status 0.}

\item{n1.cutoff}{Minimum number of samples with status 1.}

\item{study}{Name of study.}

\item{surv.outcome}{Overall survival (OS) or progression-free survival (PFS).}

\item{sig.name}{Name of signature.}

\item{method}{The default method to convert a continuous variable into a dichotomous variable is the 'median' method. The first quartile (Q1) and third quartile (Q3) can also be applied.}

\item{var.type}{If the variable is dichotomous (by default), then var.type is TRUE.}

\item{cancer.type}{Name of the cancer type for the given study.}

\item{treatment}{Name of the treatment for the given study.}
}
\description{
Fits a Cox proportional hazards regression model with dichotomous signature data using the counting process formulation of Andersen and Gill.
}
