% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSurvDicho.R
\name{geneSurvDicho}
\alias{geneSurvDicho}
\title{Fit Proportional Hazards Regression Model for Genes: Dichotomous Expression Variable}
\usage{
geneSurvDicho(
  dat.icb,
  clin = NULL,
  time.censor,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  feature,
  study,
  surv.outcome,
  n0.cutoff,
  n1.cutoff,
  method = "median",
  var.type,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{A MultiAssayExperiment (MAE) object, SummarizedExperiment (SE) object, or a data frame or matrix of gene expression data.}

\item{clin}{If dat.icb is a data frame or matrix, then it contains clinical data (as data frame or matrix). By default, it is NULL.}

\item{time.censor}{Possible censoring in months.}

\item{missing.perc}{A cutoff to remove genes with zero expression across samples.}

\item{const.int}{A pseudocount is added to the TPM (Transcripts Per Million) values before performing a log transformation.}

\item{n.cutoff}{Minimum number of samples included in the association analysis.}

\item{feature}{A vector of character strings for selected features.}

\item{study}{Name of study.}

\item{surv.outcome}{Overall survival (OS) or progression-free survival (PFS).}

\item{n0.cutoff}{Minimum number of samples with status 0.}

\item{n1.cutoff}{Minimum number of samples with status 1.}

\item{method}{The default method to convert a continuous variable into a dichotomous variable is the 'median' method. The first quartile (Q1) and third quartile (Q3) can also be applied.}

\item{var.type}{If the variable is dichotomous (by default), then var.type is TRUE.}

\item{cancer.type}{Name of the cancer type for the given study.}

\item{treatment}{Name of the treatment for the given study.}
}
\value{
A subset of results using an object of class 'coxph' representing the fit.
Outcome: Immunotherapy time-to-event outcomes including overall survival (OS) and progression-free survival (PFS).
Gene: Name of selected genes.
Study: Name of study.
Coef: Estimate of treatment effect i.e., log hazard ratio.
SE: Standard error of treatment estimate.
N: Number of samples.
Pval: Estimated p-value.
Cancer_type: A character shows the cancer type.
Treatment: A character shows the treatment type.
}
\description{
Fits a Cox proportional hazards regression model with dichotomous expression data for each gene using the counting process formulation of Andersen and Gill.
}
\examples{
Assess the association between selected genes (i.e., dichotomous variable with median cutoff) and OS in immunotherapy.
geneSurvDicho(dat.icb = ICB_Liu,
             time.censor = 36,
             missing.perc = 0.5,
             const.int = 0.001,
             n.cutoff = 15,
             feature = c('CXCL9', 'CXCL10', 'TIGIT', 
                         'CD83', 'STAT1', 'CXCL11',
                         'CXCL13', 'CD8A', 'CTLA4'),
             study = 'ICB_Liu',
             surv.outcome = 'OS',
             n0.cutoff = 5, 
             n1.cutoff = 5, 
             method = "median",
             var.type = FALSE,
             cancer.type = 'Melanoma',
             treatment = 'PD-1/PD-L1')
             
}
