% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSurvDicho.R
\name{geneSurvDicho}
\alias{geneSurvDicho}
\title{Title}
\usage{
geneSurvDicho(
  dat.icb,
  clin = NULL,
  time.censor,
  missing.perc,
  const.int = 0.001,
  n.cutoff,
  feature,
  study,
  surv.outcome,
  n0.cutoff,
  n1.cutoff,
  method = "median",
  var.type,
  cancer.type,
  treatment
)
}
\arguments{
\item{dat.icb}{aaaaa}

\item{clin}{bbbbb}

\item{time.censor}{ccccc}

\item{missing.perc}{ddddd}

\item{const.int}{eeeee}

\item{n.cutoff}{ffff}

\item{feature}{ggggg}

\item{study}{hhhh}

\item{surv.outcome}{iiii}

\item{n0.cutoff}{jjjj}

\item{n1.cutoff}{kkkk}

\item{method}{llll}

\item{var.type}{mmm}

\item{cancer.type}{nnnn}

\item{treatment}{oooo}
}
\value{
pppp
}
\description{
Title
}
