% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPSfun.R
\name{IPSfun}
\alias{IPSfun}
\title{IPS Function}
\usage{
IPSfun(expr, sig)
}
\arguments{
\item{expr}{A numeric matrix of expression data.}

\item{sig}{A data frame of list of genes.}
}
\description{
IPS Function
}
