% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMPlot.R
\name{KMPlot}
\alias{KMPlot}
\title{Function to Plot Kaplan-Meier Survival Curves}
\usage{
KMPlot(
  status,
  time,
  time.censor,
  var,
  title,
  xlab,
  ylab,
  method = "median",
  n0.cutoff,
  n1.cutoff,
  var.type = TRUE
)
}
\arguments{
\item{status}{A vector of 0 and 1, where 0 indicates 'sample was censored at time t' and 1 indicates 'sample had an event at time t'.}

\item{time}{A vector of time, in months, until endpoint or last follow-up. It is the follow up time (used with right censored data).}

\item{time.censor}{Possible censoring in months.}

\item{var}{A vector of dichotomous or continuous expression data.}

\item{title}{Label for title of the Kaplan Meier's plot.}

\item{xlab}{Label for x-axis of the Kaplan Meier's plot.}

\item{ylab}{Label for y-axis of the Kaplan Meier's plot.}

\item{method}{The default method to convert a continuous variable into a dichotomous variable is the 'median' method. The first quartile (Q1) and third quartile (Q3) can also be applied.}

\item{n0.cutoff}{Minimum number of samples with status 0.}

\item{n1.cutoff}{Minimum number of samples with status 1.}

\item{var.type}{If the variable is dichotomous (by default), then var.type is TRUE.}
}
\description{
Function to plot several Kaplan-Meier survival curves with number of individuals at risk at some time points.
}
\details{
The original version of this function was kindly provided by Dr Christos Hatzis (January, 17th 2006).
}
\examples{
Assess the association between CXCL9 (i.e., dichotomous varibale with median cutoff) and PFS in immunotherapy.
expr <- assay(ICB_small_Liu);
clin <- colData(ICB_small_Liu) \%>\% as.data.frame();
KMPlot(status = clin$event_occurred_pfs,
       time = clin$survival_time_pfs,
       time.censor = 36,
       var =  as.numeric(expr['CXCL9', ]), 
       title = "CXCL9 and PFS Association",
       xlab = "Time (Months)",
       ylab = "Progression-free Survival", 
       method = "median",
       n0.cutoff = 5, 
       n1.cutoff = 5,
       var.type = FALSE)
}
