% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafun.R
\name{metafun}
\alias{metafun}
\title{Meta-analysis via Generic Inverse Variance Method}
\usage{
metafun(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  cancer.spec = FALSE,
  treatment.spec = FALSE,
  feature
)
}
\arguments{
\item{coef}{A numeric vector of estimates of treatment effect, e.g., log hazard ratio or log odds ratio.}

\item{se}{A numeric vector of standard errors of treatment estimate.}

\item{study}{A vector of characters showing the studies' names.}

\item{pval}{A numeric vector of estimated p-values.}

\item{n}{A numeric vector of number of samples.}

\item{cancer.type}{A vector of characters showing the cancer types.}

\item{treatment}{A vector of characters showing the treatment types.}

\item{cancer.spec}{By default, it is FALSE.}

\item{treatment.spec}{By default, it is FALSE.}

\item{feature}{Name of feature.}
}
\description{
Random effect meta-analysis based on estimates (e.g. log hazard ratios) and their standard errors. The inverse variance method is used for pooling.
}
