% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPerTreatmentfun.R
\name{metaPerTreatmentfun}
\alias{metaPerTreatmentfun}
\title{Meta-analysis via Generic Inverse Variance Method: Treatment-specific Study}
\usage{
metaPerTreatmentfun(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  treatment.spec = TRUE,
  feature
)
}
\arguments{
\item{coef}{A numeric vector of estimates of treatment effect, e.g., log hazard ratio or log odds ratio.}

\item{se}{A numeric vector of standard errors of treatment estimate.}

\item{study}{A vector of characters showing the studies' names.}

\item{pval}{A numeric vector of estimated p-values.}

\item{n}{A numeric vector of number of samples.}

\item{cancer.type}{A vector of characters showing the cancer types.}

\item{treatment}{A vector of characters showing the treatment types.}

\item{treatment.spec}{By default, it is TRUE.}

\item{feature}{Name of feature.}

\item{cancer.spec}{By default, it is FALSE.}
}
\description{
Random effect meta-analysis based on estimates (e.g. log hazard ratios) and their standard errors. The inverse variance method is used for pooling.
}
