% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPerTreatmentfun.R
\name{metaPerTreatmentfun}
\alias{metaPerTreatmentfun}
\title{Title}
\usage{
metaPerTreatmentfun(
  coef,
  se,
  study,
  pval,
  n,
  cancer.type,
  treatment,
  treatment.spec = TRUE,
  feature
)
}
\arguments{
\item{coef}{}

\item{se}{}

\item{study}{}

\item{pval}{}

\item{n}{}

\item{cancer.type}{}

\item{treatment}{}

\item{treatment.spec}{}

\item{feature}{}
}
\description{
Title
}
