% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCont.R
\name{survCont}
\alias{survCont}
\title{Fit Proportional Hazards Regression Model: Continuous Expression Variable}
\usage{
survCont(status, time, time.censor, var)
}
\arguments{
\item{status}{A vector of 0 and 1, where 0 indicates 'sample was censored at time t' and 1 indicates 'sample had an event at time t'.}

\item{time}{A vector of time, in months, until endpoint or last follow-up. It is the follow up time (used with right censored data).}

\item{time.censor}{Possible censoring in months.}

\item{var}{A vector of continuous expression data.}

\item{HR}{Estimate of treatment effect i.e., log hazard ratio.}

\item{SE}{Standard error of treatment estimate.}

\item{N}{Number of samples.}

\item{Low}{Lower bound of the 95\% confidence interval.}

\item{Up}{Upper bound of the 95\% confidence interval.}

\item{Pval}{Estimated p-value.}
}
\value{
A subset of results using an object of class 'coxph' representing the fit.
}
\description{
Fits a Cox proportional hazards regression model with continuous expression data using the counting process formulation of Andersen and Gill.
}
\examples{
Assess the association between CXCL9 and OS in immunotherapy.
expr <- assay(ICB_Liu);
clin <- colData(ICB_Liu) \%>\% as.data.frame();
survCont(status = clin$event_occurred_os,
         time = clin$survival_time_os,
         time.censor = 36, 
         var = as.numeric(expr["CXCL9", ])
         )
}
