% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survCont.R
\name{survCont}
\alias{survCont}
\title{Fit Proportional Hazards Regression Model: Continuous Expression Variable}
\usage{
survCont(status, time, time.censor, var)
}
\arguments{
\item{status}{A vector of 0 and 1, where 0 indicates 'sample was censored at time t' and 1 indicates 'sample had an event at time t'.}

\item{time}{A vector of time, in months, until endpoint or last follow-up.}

\item{time.censor}{Possible censoring in months.}

\item{var}{A vector of continuous expression data.}
}
\value{
A subset of results using an object of class 'coxph' representing the fit.
}
\description{
Fits a Cox proportional hazards regression model with continuous expression data using the counting process formulation of Andersen and Gill.
}
