% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survDicho.R
\name{survDicho}
\alias{survDicho}
\title{Fit Proportional Hazards Regression Model: Dichotomous Expression Variable}
\usage{
survDicho(
  status,
  time,
  time.censor,
  var,
  n0.cutoff,
  n1.cutoff,
  method = "median",
  var.type = TRUE
)
}
\arguments{
\item{status}{A vector of 0 and 1, where 0 indicates 'sample was censored at time t' and 1 indicates 'sample had an event at time t'.}

\item{time}{A vector of time, in months, until endpoint or last follow-up. It is the follow up time (used with right censored data).}

\item{time.censor}{Possible censoring in months.}

\item{var}{A vector of dichotomous or continuous expression data.}

\item{n0.cutoff}{Minimum number of samples with status 0.}

\item{n1.cutoff}{Minimum number of samples with status 1.}

\item{method}{The default method to convert a continuous variable into a dichotomous variable is the 'median' method. The first quartile (Q1) and third quartile (Q3) can also be applied.}

\item{var.type}{If the variable is dichotomous (by default), then var.type is TRUE.}
}
\value{
A subset of results using an object of class 'coxph' representing the fit.
HR: Estimate of treatment effect i.e., log hazard ratio.
SE: Standard error of treatment estimate.
N: Number of samples.
Low: Lower bound of the 95\% confidence interval.
Up: Upper bound of the 95\% confidence interval.
Pval: Estimated p-value.
}
\description{
Fits a Cox proportional hazards regression model with dichotomous expression data using the counting process formulation of Andersen and Gill.
}
\examples{
Assess the association between CXCL9 (i.e., dichotomous variable with median cutoff) and OS in immunotherapy.
expr <- assay(ICB_Liu);
clin <- colData(ICB_Liu) \%>\% as.data.frame();
survDicho( status = clin$event_occurred_os ,
           time = clin$survival_time_os,
           time.censor= 36,
           var = as.numeric(expr['CXCL9', ]),
           n0.cutoff = 5,
           n1.cutoff = 5,
           method = "median",
           var.type = FALSE )

}
