% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Title}
\usage{
volcanoPlot(
  feature,
  coef,
  pval,
  padj,
  pos.cutoff,
  neg.cutoff,
  x.lab,
  padj.label,
  cutoff
)
}
\arguments{
\item{feature}{aaaaa}

\item{coef}{bbbbbb}

\item{pval}{ccccc}

\item{padj}{ddddd}

\item{pos.cutoff}{eeeee}

\item{neg.cutoff}{fffff}

\item{x.lab}{gggg}

\item{padj.label}{hhhh}

\item{cutoff}{iiiii}
}
\value{
jjjj
}
\description{
Title
}
